refactor: change hardcoded data to prop
